moved login in navbar to end
