+++ client/BookApp-React/components/Navbar.jsx
@@ -8,9 +8,9 @@ export default function Navbar() {
       <header>
         <nav>
           <Link to="viewbooks">Books</Link>
-          <Link to="login">Login</Link>
           <Link to="viewauthors">Authors</Link>
           <Link to="createauthor">Create Author</Link>
+          <Link to="login">Login</Link>
         </nav>
       </header>
     </div>
